+++ seldon-request-logger/app/cifardata.json
