+++ seldon-request-logger/app/test_cifar.sh
